requirements.txt
flask
flask_cors
supabase
python-dotenv
werkzeug
PyJWT